% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{accident_2013}
\alias{accident_2013}
\title{Number of fatalities in 30202 observations of 50 variables in 2013.}
\format{
A data frame with 30202 rows and 50 variables:
\describe{
\item{STATE}{State code in US}
\item{ST_CASE}{State case number}
...
}
}
\usage{
accident_2013
}
\description{
A dataset containing the locations and other attributes of 30202 crashes.
}
\keyword{datasets}
