% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_map_state}
\alias{fars_map_state}
\title{fars_map_state}
\usage{
fars_map_state(state.num, year)
}
\arguments{
\item{state.num}{Integer representation of a state.}

\item{year}{4-digit year.}
}
\description{
Plot the fars data or the state (input) from the input data of a selecter year.
}
\examples{
fars_map_state(1, 2014)

\dontrun{
fars_map_state(3, 2019)   # error
}

}
