% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_read}
\alias{fars_read}
\title{fars_read}
\usage{
fars_read(filename)
}
\arguments{
\item{filename}{File name of the csv-file.}
}
\value{
A dataframe.
}
\description{
Read a csv-data file. If file exists, the data will be loaded and a fraction of it will be shown.
}
\examples{
fars_read("accident_2013.csv.bz2")

\dontrun{
fars_read(filename = 'filedoesnotexist')
}

}
