% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_read_years}
\alias{fars_read_years}
\title{fars_read_years}
\usage{
fars_read_years(years)
}
\arguments{
\item{years}{List of years.}
}
\value{
Display the dateframe representation of the input data.
}
\description{
Read multiple CSV-files for the corresponding years of "accident_####.csv.bz2".
}
\examples{
fars_read_years(c(2013, 2014, 2015))

\dontrun{
fars_read_years(years = 2000) # error
}

}
