% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_summarize_years}
\alias{fars_summarize_years}
\title{fars_summarize_years}
\usage{
fars_summarize_years(years)
}
\arguments{
\item{years}{List of years.}
}
\value{
Display the data characteristics of the years of the corresponding CSV-file.
}
\description{
Produce a summary of the FARS Files.
}
\examples{
fars_summarize_years(c(2013, 2014, 2015))

\dontrun{
fars_summarize_years(years = 2018)
}

}
