% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_map_state.R
\name{fars_map_state}
\alias{fars_map_state}
\title{Plots the fatal accidents by location on the state}
\usage{
fars_map_state(state.num, year)
}
\arguments{
\item{state}{number and year}
}
\value{
plot of fatalities for the given state and year
if the state number is invalid it gives error message
"invalid STATE number: "

@example
fars_map_state(25,2013)
}
\description{
The fars_map_state function maps the fatality of the state
for a given year
fars_map_state function takes the state number and year as input
and maps the fatality data based on the location
}
