% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_read.R
\name{fars_read}
\alias{fars_read}
\title{read in a csv file as a tibble "data"}
\usage{
fars_read(filename)
}
\arguments{
\item{filename}{Name of the csv file as a string.}
}
\value{
tibble data frame
}
\description{
The fars_read function wraps the read_csv function.
This function takes a file in csv format reads it and returns
a tibble data frame if the file exists. Otherwise the error
message filename does not exist will be displayed.
}
\examples{
fars_read("accident_2013.csv.bz2")

}
