% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_read_years.R
\name{fars_read_years}
\alias{fars_read_years}
\title{Reads the years and gives fatalities by month and year}
\usage{
fars_read_years(years)
}
\arguments{
\item{vector}{of years}
}
\value{
lists of tibble dataframes selecting the fatal accident
data for each month in the given years. If the year is not matching
it gives the warning 'invalid year'
}
\description{
The function fars_read_years list the fatality data for each
month in the years given
fars_read_years function takes vector of years as input
and selects the MONTH and year columns and lists tibble data
frame for each year
}
\examples{
fars_read_years(2013)

}
