% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_summarize_years.R
\name{fars_summarize_years}
\alias{fars_summarize_years}
\title{Summarises the fatalities by each month of the years}
\usage{
fars_summarize_years(years)
}
\arguments{
\item{vector}{of years}
}
\value{
the summary of fatalities by month and year
}
\description{
The fars_summarize_years function gives the total number
of fatalities for each month of given years as input
fars_summarize_years summarises the fatalities each month for
all the given years
}
