% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_filename.R
\name{make_filename}
\alias{make_filename}
\title{Makes a filename formatted with the given year input}
\usage{
make_filename(year)
}
\arguments{
\item{year}{as a string.}
}
\value{
file name with the year attached
}
\description{
The function make_filename creates a filename
with the given year make_filename takes string year as input and appends that year to the
filename
}
\examples{
make_filename("2013")

}
